requirements.txt
fastapi
uvicorn
dnspython
python-whois
httpx
beautifulsoup4
pandas
